requirements.txt
Flask==2.1.2
speedtest-cli==2.1.3
